requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
pydantic
demucs
aiofiles
soundfile
google-auth
